feature: item tree state reactive to the view refactor: refactoring various components refactor: getting rod of debugging code
