+++ src/actions/action-handlers.js
@@ -16,7 +16,23 @@ export function keywordsHandler(currentState, action) {
 }
 
 export function itemTreeHandler(currentState, action) {
-  return action.type === 'newItemTree' ?
-      {...currentState, itemTree: action.value} :
-      currentState;
+  switch (action.type) {
+    case 'newItemTree':
+      return {
+        ...currentState,
+        itemTree: action.value,
+      };
+    case 'newCaseTree':
+      return {
+        ...currentState,
+        caseTree: action.value,
+      };
+    case 'selectedItemTree':
+      return {
+        ...currentState,
+        selectedItem: action.value,
+      };
+    default:
+      return currentState;
+  }
 }
