+++ src/components/sidebar.css
@@ -1,6 +1,6 @@
 .sidebar {
     width: 55%;
-    background-color: #f1f1f1;
+    background-color: #505050;
     padding: 20px;
     height: 100%;
     overflow: auto;
