+++ src/components/sidebar.html
@@ -1,21 +1,26 @@
 <aside class="sidebar">
   <dl>
     <template repeat.for="item of itemTree & state">
-      <dt>
+      <dt data-key.bind="item.key">
         <a href="#" click.trigger="toggleUnit('level0', $index)">+</a>
-        <a href="#" click.trigger="selectUnit('blias')">${item?.bailiffName}</a>
+        <a href="#" click.trigger="selectUnit(item.key)">${item?.bailiffName}</a>
         <span class="badge">${treeState.level0.item[$index].expanded} index is ${$index}</span>
       </dt>
       <dd repeat.for="person of item?.debtors" show.bind="treeState.level0['item'][$index].expanded">
         <ul>
           <li>
             <a href="#" click.trigger="toggleUnit('level1', $index)">+</a>
-            <a href="#" click.trigger="selectUnit('blias')">${person?.name}</a>
+            <a
+                    href="#"
+                    data-work-unit-id.bind="person.workUnitId"
+                    class.two-way="selectedUnit === person.workUnitId ? 'selected' : ''"
+                    click.trigger="selectUnit(person.workUnitId)">
+              ${person?.name}
+            </a>
             <span class="badge">${treeState.level1.item[$index].expanded} index is ${$index}</span>
             <ul repeat.for="case of item?.cases" show.bind="treeState.level1['item'][$index].expanded">
                 <li>
-                    <a href="#" click.trigger="toggleUnit('level1', $index)">+</a>
-                    <a href="#" click.trigger="selectUnit('blias')">${case?.package}</a>
+                    <a href="#" data-work-unit-id.bind="case.workUnitId" click.trigger="selectUnit(case.workUnitId)">${case?.package}</a>
                     <span class="badge">${treeState.level2.item[$index].expanded} index is ${$index}</span>
                 </li>
             </ul>
@@ -24,8 +29,4 @@
       </dd>
     </template>
   </dl>
-  <code>
-    ${itemTree[0]?.bailiffName & state}
-    ${keywords & state}
-</code>
 </aside>
