+++ src/constants/initialstate.js
@@ -2,4 +2,6 @@ export const initialState = {
   keywords: '',
   items: [],
   itemTree: [],
+  caseTree: [],
+  selectedItem: null,
 };
