+++ src/mock-server/getPersonAddresses.json
@@ -1,0 +1,18 @@
+{
+  "addresses": [
+    {
+      "id": "c8c9772d-4c6b-6865-e053-0100007f6ed2",
+      "settlementName": "ПРОВАДИЯ",
+      "zipCode": "9201",
+      "address": "Ж.К.СЕВon RpL25 uvog tiJ3 rUo72",
+      "dateAdded": "2011-06-07T13:39:52+03:00"
+    },
+    {
+      "id": "8e1f0c1f-0efd-4b00-9d8e-413b9ecae3d6",
+      "contactActivityTypeId": null,
+      "settlementName": "МАЛО СЕЛО",
+      "zipCode": "2636",
+      "address": "ул.Първа 6"
+    }
+  ]
+}
