+++ test/my-app.spec.js
@@ -1,5 +1,5 @@
 import {render} from './helper';
-import {MyApp} from '../src/my-app';
+import {MyApp} from '../src/components/my-app';
 
 describe('my-app', () => {
   it('should render message', async () => {
